LayoutTests/editing/deleting/delete-across-editable-content-boundaries-3-expected.txt

This tests to see that selected editable nodes that are not in both the start and end editable roots are cleared.  You should see one non-editable piece and an editable one and then two empty table cells.  The table cells are editable but are preserved and not removed because they are in a non-editable region.  They should be completely cleared out, even the second one that contains a non-editable piece. <radr://problem/5026848>
| "\n"
| <table>
|   border="1"
|   contenteditable="false"
|   "\n    "
|   <tbody>
|     <tr>
|       "\n        "
|       <td>
|         "Non-editable"
|       "\n        "
|       <td>
|         contenteditable="true"
|         "Editable <#selection-caret>content"
|       "\n        "
|       <!--  The following cell should be removed, even though it contains non-editable pieces.  -->
|       "\n        "
|       <td>
|         contenteditable="true"
|       "\n        "
|       <td>
|         contenteditable="true"
|       "\n    "
|     "\n"
